+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/EvaluableLazyParameterCreator.scala
@@ -1,14 +1,14 @@
 package pl.touk.nussknacker.engine.compile.nodecompilation
 
+import cats.implicits._
 import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.LazyParameter._
 import pl.touk.nussknacker.engine.api.context.ValidationContext
-import pl.touk.nussknacker.engine.api.parameter.ParameterName
+import pl.touk.nussknacker.engine.api.definition.Validator
 import pl.touk.nussknacker.engine.api.typed.typing._
 import pl.touk.nussknacker.engine.compile.ExpressionCompiler
-import pl.touk.nussknacker.engine.compiledgraph.BaseCompiledParameter
+import pl.touk.nussknacker.engine.compiledgraph.CompiledParameter
 import pl.touk.nussknacker.engine.expression.ExpressionEvaluator
-import pl.touk.nussknacker.engine.expression.parse.CompiledExpression
 import pl.touk.nussknacker.engine.graph.expression.Expression
 
 // This class looks like a LazyParameter but actually it's not - it's a creator of the LazyParameter.
@@ -33,17 +33,16 @@ final class EvaluableLazyParameterCreator[T <: AnyRef](
   }
 
   private def createEvaluableLazyParameter(deps: EvaluableLazyParameterCreatorDeps) = {
+    implicit val implicitNodeId: NodeId   = nodeId
+    implicit val implicitJobData: JobData = deps.jobData
     val compiledExpression = deps.expressionCompiler
       .compile(expression, Some(parameterDef.name), inputValidationContext, parameterDef.typ)(nodeId)
       .valueOr(err =>
         throw new IllegalArgumentException(s"Compilation failed with errors: ${err.toList.mkString(", ")}")
       )
-    val compiledParameter: BaseCompiledParameter = new BaseCompiledParameter {
-      override val name: ParameterName                      = parameterDef.name
-      override val expression: CompiledExpression           = compiledExpression.expression
-      override val shouldBeWrappedWithScalaOption: Boolean  = parameterDef.scalaOptionParameter
-      override val shouldBeWrappedWithJavaOptional: Boolean = parameterDef.javaOptionalParameter
-    }
+    val compiledValidators: List[Validator] =
+      deps.expressionCompiler.compileValidatorsOrThrow(parameterDef, inputValidationContext.globalVariables)
+    val compiledParameter = CompiledParameter(compiledExpression, parameterDef, compiledValidators)
     new EvaluableLazyParameter[T](
       compiledParameter,
       deps.expressionEvaluator,
@@ -58,7 +57,7 @@ final class EvaluableLazyParameterCreator[T <: AnyRef](
 class EvaluableLazyParameterCreatorDeps(
     val expressionCompiler: ExpressionCompiler,
     val expressionEvaluator: ExpressionEvaluator,
-    val jobData: JobData
+    val jobData: JobData,
 )
 
 class DefaultToEvaluateFunctionConverter(deps: EvaluableLazyParameterCreatorDeps) extends ToEvaluateFunctionConverter {